feat(rust): merge hello world and DynamoDB templates
