--- rust/cookiecutter-aws-sam-hello-rust-dynamodb/cookiecutter.json
@@ -1,5 +1,0 @@
-{
-    "project_name": "My Project",
-    "project_slug": "{{ cookiecutter.project_name.lower().replace(' ', '-') }}",
-    "architecture": ["x86_64", "arm64"]
-  }
