--- rust/cookiecutter-aws-sam-hello-rust-dynamodb/{{ cookiecutter.project_slug }}/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "{{ cookiecutter.project_slug }}"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-aws-config = "0.4"
-aws-sdk-dynamodb = "0.4"
-aws-smithy-client = { version = "0.34", features = ["test-util"] }
-lambda_http = "0.4"
-tokio = { version = "1", features = ["full"] }
-
-[dev-dependencies]
-aws-smithy-http = "0.34"
-aws-types = { version = "0.4", features = ["hardcoded-credentials"] }
-http = "0.2"
