+++ rust/cookiecutter-aws-sam-hello-rust/{{ cookiecutter.project_slug }}/Cargo.toml
@@ -6,5 +6,13 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+aws-config = "0.4"
+aws-sdk-dynamodb = "0.4"
+aws-smithy-client = { version = "0.34", features = ["test-util"] }
 lambda_http = "0.4"
 tokio = { version = "1", features = ["full"] }
+
+[dev-dependencies]
+aws-smithy-http = "0.34"
+aws-types = { version = "0.4", features = ["hardcoded-credentials"] }
+http = "0.2"
